chore(package.json): add react-player dependency refactor(Clip.jsx): remove unused code and replace iframe with ReactPlayer component feat(Container.jsx): create Container component to fetch and display videos refactor(Main.jsx): replace Clip component with Container component
