+++ client/package.json
@@ -9,6 +9,7 @@
     "bootstrap": "^5.1.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-player": "^2.12.0",
     "react-scripts": "^5.0.0",
     "web-vitals": "^1.1.2"
   },
